Improved the reading of the storage files:
There are two new Exceptions:
1. EndOfBlockException if the control word that marks the end of a data bock is missing.
2. StartOfBlockException if the control word that marks the begin of a data block is missing.

The Map Creation Exception has a more detailed message by overriding the getMessage method.

The StorageReader lost some unnecessary code.
If the end of a data block is not checked the next time calling nextString will result in an EndOfBlockException!

The MapCreator lost the unnecessary while-loop and uses the new exceptions.