All objects should be connected now. Rename objects.card Jail to objects.card Arrest Move MoreThanOneInstanceException to objects.exceptions.data Rename MoreThanOneInstanceException -> MoreThanOneDataSetException
Connector:
Add method to connect Cards with needed Objects.
Exception:
Add new Exception if the Object needed for a card is not found bay name.
CardStack:
Add toArray method that returns the Card array.
Arrest:
Add attribute and setter for the Field jail.
GoTo:
Add attribute, setter and getter for the Field they have to point.
PayFineTakeCard:
Add attribute, setter and getter to the community CardStack.